pyproject.toml
[tool.poetry]
name = "poster"
version = "0.1.0"
description = ""
authors = ["Lukas Boehm <lukas.l.boehm@fau.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-fire = "^0.1.0"
wandb = "^0.18.1"
submitit = "^1.5.1"
hydra-submitit-launcher = "^1.2.0"
hydra-zen = "^0.13.0"
typer-slim = "^0.12.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
